multiplies two positive numbers

+++ 0x0C-more_malloc_free/101-mul.c
@@ -10,7 +10,7 @@ int main(int argc, char *argv[])
 {
 
 	unsigned long int mul;
-	int num1, num2;
+	unsigned long int num1, num2;
 
 
 	if (argc != 3)
